internal_documents_core/pyproject.toml
[project]
name = "internal-documents-core"
version = "0.1.0"
description = "Internal document ingestion and vector store population"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.0.0,<3.0.0",
    "google-auth-httplib2>=0.2.0,<1.0.0",
    "google-auth-oauthlib>=1.2.0,<2.0.0",
    "openai>=1.0.0,<2.0.0",
    "psycopg[binary]>=3.2.0,<4.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "tqdm>=4.66.0,<5.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0,<9.0.0",
]

[tool.setuptools.packages.find]
include = ["core*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
